+++ src/app/admin/page.tsx
@@ -1,4 +1,4 @@
-import ButtonAppBar from '@/components/buttonAppBar';
+import NavBar from '@/components/NavBar';
 import DataTable from '@/components/dashboard';
 import DashboardMenu from '@/components/dashboardMenu';
 import styles from '../page.module.css';
@@ -6,7 +6,7 @@ import styles from '../page.module.css';
 export default async function Home() {
   return (
     <main className={styles['main']}>
-      <ButtonAppBar />
+      <NavBar />
       <DashboardMenu />
       <DataTable />
     </main>
