+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Menu from '@/components/menu';
+import Menu from '@/components/Menu';
 
 export default function Home() {
   return <Menu />;
